move RenameElementConfig to event-types
