+++ src/app/modules/event-types-in.ts
@@ -2,7 +2,7 @@
 
 
 // request to let the user create a new node in the context of the currently selected tree node
-// payload NewElementConfig
+// payload: NewElementConfig
 export interface NewElementConfig {
   indent: boolean;
   validateName: (newName: string) => { valid: boolean, message?: string };
@@ -12,4 +12,9 @@ export interface NewElementConfig {
 export const TREE_NODE_CREATE_AT_SELECTED = 'treenode.create-at-selected';
 
 // request to let the user rename the currently selected tree node
+// payload: RenameElementConfig
+export interface RenameElementConfig {
+  validateName: (newName: string) => { valid: boolean, message?: string };
+  renameElement: (newName: string) => boolean;
+}
 export const TREE_NODE_RENAME_SELECTED = 'treenode.rename-selected';
